+++ vllm/lora/layers/base.py
@@ -1,20 +1,16 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
-
-
 from typing import TYPE_CHECKING, Optional, Union
 
 import torch
 import torch.nn as nn
-
 from transformers import PretrainedConfig
 
 from vllm.config import LoRAConfig
 
 if TYPE_CHECKING:
     from vllm.lora.punica_wrapper import PunicaWrapperBase
-
 
 
 class BaseLayerWithLoRA(nn.Module):
@@ -71,6 +67,3 @@ class BaseLayerWithLoRA(nn.Module):
     ) -> bool:
         """Returns True if the layer can be replaced by this LoRA layer."""
         raise NotImplementedError
-
-
-
