+++ vllm/lora/layers/qkv_x_parallel_linear.py
@@ -1,5 +1,8 @@
-
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 from .base import BaseLayerWithLoRA
+
+
 #TODO: Implement this
 class QKVCrossParallelLinearWithLoRA(BaseLayerWithLoRA):
     pass
